# api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "CloudShell Sandbox API",
    "description": "",
    "version": "1.0.0"
  },
  "produces": [
    "application/json"
  ],
  "host": "ec2-52-9-19-65.us-west-1.compute.amazonaws.com:8080",
  "basePath": "/api",
  "tags": [
    {
      "name": "sandbox",
      "description": "'Sandboxes are self contained, isolated, interactive environments containing networked physical and virtual devices and apps. A sandbox is includes orchestration commands for interacting with the sandbox and its components. Users can interact with a sandbox to experiment, test and develop'"
    },
    {
      "name": "blueprint",
      "description": "Blueprints define a template for a sandbox which contains its basic definitions, capabilities and structure. They are made available to users as a self service catalog."
    }
  ],
  "paths": {

    "/v1/blueprints/{blueprint_identifier}/setup_sandbox": {
      "post": {
        "summary": "",
        "description": "Create a new sandbox instance from a given blueprint.",
        "x-swagger-router-controller": "Sandbox",
        "tags": [
          "sandbox"
        ],
        "operationId": "v1BlueprintsBlueprintIdentifierSetupSandboxPost",
        "parameters": [
          {
            "name": "blueprint_identifier",
            "in": "path",
            "description": "the name or id of the blueprint",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Sandbox scheduling parameters",
            "required": false,
            "schema": {
              "$ref": "#/definitions/SandboxStartParameters"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "$ref": "#/definitions/Sandbox"
            },
            "examples": {
              "application/json": {
                "data": {
                  "id": "4d9f97c2-7f0d-411d-9501-e9c5de41b4e4",
                  "name": "APIC-EM 1.0GA DataCenter Discovery",
                  "category": "APIC-EM",
                  "availability": "available",
                  "description": "An example discovery application intended for develoeprs to experiment with",
                  "_links": [
                    {
                      "self": "/v1/blueprints/4d9f97c2-7f0d-411d-9501-e9c5de41b4e4"
                    },
                    {
                      "create_sandbox": "/v1/blueprints/4d9f97c2-7f0d-411d-9501-e9c5de41b4e4/create_sandbox"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/sandboxes/{sandbox_id}": {
      "get": {
        "summary": "",
        "description": "Returns detailed information for a specific sandbox",
        "x-swagger-router-controller": "Sandbox",
        "tags": [
          "sandbox"
        ],
        "operationId": "v1SandboxesSandboxIdGet",
        "parameters": [
          {
            "name": "sandbox_id",
            "in": "path",
            "description": "Get the detailed information for a specific sandbox.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Sandbox"
              }
            },
            "examples": {
              "application/json": {
                "data": {
                  "id": "4d9f97c2-7f0d-411d-9501-e9c5de41b4e4",
                  "name": "CI Build 123123 APC-IM Test Policy",
                  "state": "in setup",
                  "description": "An example discovery application intended for develoeprs to experiment with",
                  "components": [
                    {
                      "name": "Client Desktop",
                      "description": "Represents a client machine connected to a WAN network",
                      "type": "SandboxComponent",
                      "component_type": "Virtual Machine",
                      "connection_interfaces": [
                        {
                          "name": "Remote Desktop",
                          "type": "ConnectionInterface",
                          "connection_interface_type": "WebRemoteLink",
                          "url": "http://sandbox_api_remote_server.com/nav?id=c%2FDEFAULT&amp;guac.hostname=192.33.45.66&amp;guac.protocol=rdp&amp;guac.port=3389&amp;guac.username=test&amp;guac.password=D*SDAD&"
                        }
                      ]
                    }
                  ],
                  "attributes": [
                    {
                      "name": "Version",
                      "type": "ComponentAttribute",
                      "value": "2.3.0"
                    }
                  ],
                  "_links": [
                    {
                      "self": "/v1/sanboxes/4d9f97c2-7f0d-411d-9501-e9c5de41b4e4"
                    },
                    {
                      "end": "/v1/sanboxes/4d9f97c2-7f0d-411d-9501-e9c5de41b4e4/end"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/v1/sandboxes": {
      "get": {
        "summary": "",
        "description": "List the active and scheduled sandboxes. Optionally also show historic sandboxes.",
        "x-swagger-router-controller": "Sandbox",
        "tags": [
          "sandbox"
        ],
        "operationId": "v1SandboxesGet",
        "parameters": [
          {
            "name": "show_historic",
            "in": "query",
            "description": "Also show historic (ended) sandboxes. Default false.",
            "required": false,
            "type": "boolean"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/SandboxDescription"
              }
            },
            "examples": {
              "application/json": {
                "data": [
                  {
                    "id": "4d9f97c2-7f0d-411d-9501-e9c5de41b4e4",
                    "name": "My Sandbox For DataCenter Discovery",
                    "state": "ready",
                    "description": "An example discovery application intended for develoeprs to experiment with",
                    "_links": [
                      {
                        "self": "/v1/sanboxes/4d9f97c2-7f0d-411d-9501-e9c5de41b4e4"
                      },
                      {
                        "end": "/v1/sanboxes/4d9f97c2-7f0d-411d-9501-e9c5de41b4e4/end"
                      }
                    ]
                  }
                ]
              }
            }
          }
        }
      }
    },
    "/v1/sandboxes/{sandbox_id}/teardown_sandbox": {
      "post": {
        "summary": "",
        "description": "End an active sandbox, starting its teardown process.",
        "x-swagger-router-controller": "Sandbox",
        "tags": [
          "sandbox"
        ],
        "operationId": "v1SandboxesSandboxIdTeardownSandboxPost",
        "parameters": [
          {
            "name": "sandbox_id",
            "in": "path",
            "description": "Sandbox id .",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      }
    },
    "/v1/blueprints": {
      "get": {
        "summary": "",
        "description": "Returns the blueprints available in the catalog.",
        "x-swagger-router-controller": "Blueprint",
        "tags": [
          "blueprint"
        ],
        "operationId": "v1BlueprintsGet",
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/BlueprintDescription"
              }
            },
            "examples": {
              "application/json": {
                "data": [
                  {
                    "id": "4d9f97c2-7f0d-411d-9501-e9c5de41b4e4",
                    "type": "blueprint",
                    "name": "APIC-EM 1.0GA DataCenter Discovery",
                    "category": "APIC-EM",
                    "description": "An example discovery application intended for develoeprs to experiment with",
                    "_links": [
                      {
                        "self": "/v1/blueprints/4d9f97c2-7f0d-411d-9501-e9c5de41b4e4"
                      },
                      {
                        "create_sandbox": "/v1/blueprints/4d9f97c2-7f0d-411d-9501-e9c5de41b4e4/create_sandbox"
                      }
                    ]
                  },
                  {
                    "id": "1c197b81-4213-467b-8a7a-f51ebd33a4b2",
                    "type": "blueprint",
                    "name": "APIC-EM 1.0GA Hello world example",
                    "category": "APIC-EM",
                    "description": "An example hello world application intended for develoeprs to experiment with",
                    "_links": [
                      {
                        "self": "/v1/blueprints/1c197b81-4213-467b-8a7a-f51ebd33a4b2"
                      },
                      {
                        "create_sandbox": "/v1/blueprints/1c197b81-4213-467b-8a7a-f51ebd33a4b2/create_sandbox"
                      }
                    ]
                  }
                ]
              }
            }
          }
        }
      }
    },
    "/v1/blueprints/{blueprint_identifier}": {
      "get": {
        "summary": "",
        "description": "Get detailed information for a specifi blueprint",
        "x-swagger-router-controller": "Blueprint",
        "tags": [
          "blueprint"
        ],
        "operationId": "v1BlueprintsBlueprintIdentifierGet",
        "parameters": [
          {
            "name": "blueprint_identifier",
            "in": "path",
            "description": "the name or id of the blueprint",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "schema": {
              "$ref": "#/definitions/Blueprint"
            },
            "examples": {
              "application/json": {
                "data": {
                  "id": "4d9f97c2-7f0d-411d-9501-e9c5de41b4e4",
                  "name": "APIC-EM 1.0GA DataCenter Discovery",
                  "category": "APIC-EM",
                  "availability": "available",
                  "description": "An example discovery application intended for develoeprs to experiment with",
                  "_links": [
                    {
                      "self": "/v1/blueprints/4d9f97c2-7f0d-411d-9501-e9c5de41b4e4"
                    }
                  ],
                  "_actions": [
                    {
                      "create_sandbox": null,
                      "method": "POST",
                      "body": {
                        "$ref": "#/definitions/SandboxStartParameters"
                      },
                      "href": "/v1/blueprints/1c197b81-4213-467b-8a7a-f51ebd33a4b2/create_sandbox"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    }

  },
  "definitions": {
    "SandboxStartParameters": {
      "type": "object",
      "properties": {
        "duration": {
          "type": "string",
          "format": "datetime",
          "description": "Duration for this instance using the \"ISO 8601\" Standard. (e.g 'T23H' or 'T4H2M')"
        },
        "name": {
          "type": "string",
          "example": "APIC Enterprise for test",
          "description": "A name to identify the new blueprint. If you don't specify a name one will be generated automatically."
        }
      },
      "title": "data"
    },
    "Sandbox": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "id": {
          "type": "string",
          "format": "string",
          "example": "ef19157e-0c8e-420b-9c8b-d7e76c46ab8b"
        },
        "name": {
          "type": "string",
          "format": "string",
          "example": "APIC-EM GA"
        },
        "state": {
          "type": "string",
          "example": "available",
          "description": "current state of the sandbox. Possible values: Pending, Setup, Ready, Teardown, Ended",
          "enum": [
            "Pending",
            "Setup",
            "Ready",
            "Teardown",
            "Ended"
          ]
        },
        "components": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SandboxComponent"
          }
        },
        "apps": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SandboxApp"
          }
        }
      }
    },
    "ConnectionInterface": {
      "type": "object",
      "required": [
        "connection_interface_type",
        "type"
      ],
      "discriminator": "connection_interface_type",
      "properties": {
        "type": {
          "type": "string"
        },
        "connection_interface_type": {
          "type": "string"
        }
      }
    },
    "WebRemoteDesktopLink": {
      "allOf": [
        {
          "$ref": "#/definitions/ConnectionInterface"
        },
        {
          "type": "object",
          "required": [
            "url"
          ],
          "properties": {
            "url": {
              "type": "string",
              "description": "Navigate to this link to remote to the machine"
            }
          }
        }
      ],
      "description": "This link can be used to remote desktop over the web to the target compoenent"
    },
    "SandboxComponent": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "type": {
          "type": "string",
          "format": "string"
        },
        "name": {
          "type": "string",
          "format": "string",
          "example": "Server1"
        },
        "component_type": {
          "type": "string"
        },
        "description": {
          "type": "string",
          "format": "string"
        },
        "connection_interfaces": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ConnectionInterface"
          }
        },
        "attributes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ComponentAttribute"
          }
        }
      }
    },
    "ComponentAttribute": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "type": {
          "type": "string",
          "format": "string"
        },
        "name": {
          "type": "string",
          "format": "string",
          "example": "Version"
        },
        "value": {
          "type": "string",
          "format": "string"
        }
      }
    },
    "SandboxApp": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "type": {
          "type": "string",
          "format": "string"
        },
        "name": {
          "type": "string",
          "format": "string",
          "example": "Server1"
        },
        "description": {
          "type": "string",
          "format": "string"
        },
        "connect_address": {
          "type": "string",
          "format": "string"
        }
      }
    },
    "SandboxDescription": {
      "type": "object",
      "required": [
        "id",
        "name"
      ],
      "properties": {
        "id": {
          "type": "string",
          "format": "string",
          "example": "ef19157e-0c8e-420b-9c8b-d7e76c46ab8b"
        },
        "name": {
          "type": "string",
          "format": "string",
          "example": "APIC-EM GA"
        },
        "blueprint": {
          "$ref": "#/definitions/BlueprintReference"
        },
        "state": {
          "type": "string",
          "example": "available",
          "description": "current state of the sandbox. Possible values: Pending, Setup, Ready, Teardown, Ended",
          "enum": [
            "Pending",
            "Setup",
            "Ready",
            "Teardown",
            "Ended"
          ]
        }
      }
    },
    "BlueprintReference": {
      "type": "object",
      "required": [
        "id",
        "name"
      ],
      "properties": {
        "id": {
          "type": "string",
          "format": "string",
          "example": "ef19157e-0c8e-420b-9c8b-d7e76c46ab8b"
        },
        "name": {
          "type": "string",
          "format": "string",
          "example": "APIC-EM GA"
        }
      }
    },
    "BlueprintDescription": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "id": {
          "type": "string",
          "format": "string",
          "example": "ef19157e-0c8e-420b-9c8b-d7e76c46ab8b"
        },
        "name": {
          "type": "string",
          "format": "string",
          "example": "APIC-EM GA"
        },
        "category": {
          "type": "string",
          "format": "string",
          "example": "developers"
        },
        "description": {
          "type": "string",
          "format": "string",
          "example": "A short description about the blueprint"
        },
        "_links": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "create_sandbox": {
                "type": "string"
              },
              "self": {
                "type": "string"
              }
            }
          }
        }
      },
      "title": "data"
    },
    "Blueprint": {
      "type": "object",
      "required": [
        "name"
      ],
      "properties": {
        "id": {
          "type": "string",
          "format": "string",
          "example": "ef19157e-0c8e-420b-9c8b-d7e76c46ab8b"
        },
        "name": {
          "type": "string",
          "format": "string",
          "example": "APIC-EM GA"
        },
        "category": {
          "type": "string",
          "format": "string",
          "example": "developers"
        }
      }
    }
  }
}